requirements.txt
requests
python-dotenv
networkx
